pyproject.toml
[project]
name = "py_dst_examples"
version = "0.1.0"
description = "DST Examples"
authors = [
    { name = "Dominik Tornow", email = "dominik.tornow@gmail.com" }
]
dependencies = [
    "resonate-sdk>=0.1.13",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/py_dst_examples"]